+++ rpc/backend/node_info.go
@@ -21,6 +21,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto"
+
 	"github.com/evmos/evmos/v18/crypto/ethsecp256k1"
 	rpctypes "github.com/evmos/evmos/v18/rpc/types"
 	"github.com/evmos/evmos/v18/server/config"
@@ -31,6 +32,11 @@ import (
 // Accounts returns the list of accounts available to this node.
 func (b *Backend) Accounts() ([]common.Address, error) {
 	addresses := make([]common.Address, 0) // return [] instead of nil if empty
+
+	if !b.cfg.JSONRPC.AllowInsecureUnlock {
+		b.logger.Debug("account unlock with HTTP access is forbidden")
+		return addresses, fmt.Errorf("account unlock with HTTP access is forbidden")
+	}
 
 	infos, err := b.clientCtx.Keyring.List()
 	if err != nil {
@@ -77,6 +83,11 @@ func (b *Backend) Syncing() (interface{}, error) {
 
 // SetEtherbase sets the etherbase of the miner
 func (b *Backend) SetEtherbase(etherbase common.Address) bool {
+	if !b.cfg.JSONRPC.AllowInsecureUnlock {
+		b.logger.Debug("account unlock with HTTP access is forbidden")
+		return false
+	}
+
 	delAddr, err := b.GetCoinbase()
 	if err != nil {
 		b.logger.Debug("failed to get coinbase address", "error", err.Error())
@@ -217,6 +228,11 @@ func (b *Backend) ImportRawKey(privkey, password string) (common.Address, error)
 // ListAccounts will return a list of addresses for accounts this node manages.
 func (b *Backend) ListAccounts() ([]common.Address, error) {
 	addrs := []common.Address{}
+
+	if !b.cfg.JSONRPC.AllowInsecureUnlock {
+		b.logger.Debug("account unlock with HTTP access is forbidden")
+		return addrs, fmt.Errorf("account unlock with HTTP access is forbidden")
+	}
 
 	list, err := b.clientCtx.Keyring.List()
 	if err != nil {